Added a grid
I added the option for the grid to be shown. The grid works perfectly fine, and it doesn't break when you change zoom from 1 to a higher value. However, the alive/dead tiles slightly break when changing the zoom setting. The size of the tile is perfectly fine but the placement is slightly off.